[IMP] Improve packaging
Rename `launcher.py` to the more pythonic `__main__.py`.
This change allow you to start python on the project folder in
order to start the application. Also if someday we package the bot
as a real python package on pypi.org and we install it, we'll be
able to start the bot from anywhere using `python -m <project>`

Create the missing `__init__.py` file at the project root

Create a template for the configuration file which is used when
the config file is missing on the filesystem.

Update the `requirement.txt` file