Generic way to register hastebin-like services
The current system becomes more and more complicated as new
services get supported. The `get_raw` function is an all-in-one
function that performs hardcoded site verification and hardcoded
string manipulation (either to extract a token or to sanitize the
uri).

The new system provides a simple decorator to register pastebin
handlers. It uses the function's name as site matching. String
sanitization and token extraction is possible thanks to a regexp
that parse the uri.

The handler registration is done as follow, it takes a site name
and a regexp, it returns an absolute URL with an http(s) scheme.

	@pastebin(regexp)
	def example_com(address, match):
	    return f'https://{address.netloc}/{match.group(0)}'

Where:
* `pastebin` is the decorator
* `regexp` is the regexp used to match the url path
* `example.com` is the domain name to match
* `address` is the complete object returned by `urllib.parse.urlparse`
* `match` is the matched path object